# handlers/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Voxcina Backend API",
    "description": "REST API of the Voxcina online store (Go backend).\n\nThe storefront (`voxcina.com`) proxies every `/api/*` request to this service, so **Try it out** on this page runs against the same origin — no CORS setup needed.\n\n## Authentication\n\nTwo security schemes are used:\n\n- **`bearerAuth`** — the JWT access token issued by `POST /api/users/login` (or the OTP flows). Ordinary user endpoints accept any valid token; endpoints marked *Admin* additionally require `role == \"admin\"` and *Staff* endpoints accept `admin` or `staff` (enforced server-side on every request against the live user record).\n- **`serviceApiKey`** — the `X-API-Key` header for server-to-server bot/channel integrations under `/api/auth/external/*` (created by admins in the external-services panel).\n\nTokens are stateless but every request is re-checked against the database: logout, deactivation, role changes and password changes take effect immediately.\n\n## Response envelope\n\nSuccessful responses are bare JSON objects (or arrays) whose shape mirrors the Go models. Auth/permission failures return `{\"error\": \"...\", \"code\": \"...\"}`.\n\n> This document describes the core surfaces of the API. Field-level fidelity is best-effort; the Go models under `models/` are the source of truth.",
    "version": "1.0.0",
    "contact": {
      "name": "Voxcina",
      "url": "https://voxcina.com"
    }
  },
  "servers": [
...
  "security": [],
...
  "paths": {
...
    "/api/admin/docs/openapi.json": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "This OpenAPI document (admin only)",
        "operationId": "getOpenApiSpec",
        "description": "The JSON served here — the source of the Swagger UI at `/swagger`.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OpenAPI 3 document",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Access token from POST /api/users/login (Authorization: Bearer <token>). Admin/Staff gates check the LIVE role in MongoDB on every request."
      },
      "serviceApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "External service API key (Telegram/Bale/Instagram bots), managed in the admin external-services panel."
      }
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Missing, invalid or expired access token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "examples": {
              "expired": {
                "value": {
                  "error": "Token expired",
                  "code": "TOKEN_EXPIRED"
                }
              },
              "revoked": {
                "value": {
                  "error": "Session revoked",
                  "code": "INVALID_TOKEN"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Authenticated but not allowed (role gate)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            },
            "example": {
              "error": "Admin access required",
              "code": "INSUFFICIENT_ROLE"
            }
          }
        }
      },
...
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code (auth endpoints)"
          }
        }
      },
...
    }
  },
...
}